Done : 7_Binary_Tree

+++ 7_Binary_Tree/srcs/main.c
@@ -1,4 +1,4 @@
-#include "bintree.h"
+#include "../includes/bintree.h"
 
 /*
 ==================================================================
@@ -7,30 +7,32 @@
 
 ==================================================================
 */
+int deleteflag = 0;
 
+//인자로 주어진 position의 부모 노드를 찾아 주소를 리턴하는 함수
 BinTreeNode *find_parent(BinTree *tree,char *position)
 {
-    BinTreeNode *node = NULL;
+    BinTreeNode *node = NULL;//부모 노드의 주소 넣을 포인터
     int str_len = 0;
     int i = 0;
 
     str_len = strlen(position);
-    // printf("check position 2 - find_parent() : %s\n", position);
-    // printf("✅%d\n", str_len);
     if (str_len == 0)
     {
         printf("[error] No input in position\n");
         return (FALSE);
     }
-    //printf("position (str_len) - %d\n", str_len);
     node = tree->pRootNode;
     while (i < str_len - 1)
     {
-        //printf("position : %s    index : %d\n", position, i);
         if (position[i] == 'l' || position[i] == 'L')
+        {    
             node = node->pLeftChild;
+        }
         else if (position[i] == 'r' || position[i] == 'R')
+        {
             node = node->pRightChild;
+        }
         else
         {
             printf("[error] invalid position value.\n");
@@ -45,13 +47,12 @@ BinTreeNode *find_parent(BinTree *tree,char *position)
     }
     return (node);
 }
-
+//특정 위치의 노드 주소를 반환하는 함수.
 BinTreeNode *getNode(BinTree *tree,char *position)
 {
-    BinTreeNode *node = NULL;
-    BinTreeNode *preturn = NULL;
-    int         i = strlen(position) - 1;
-
+    BinTreeNode *node = NULL;//부모 노드의 주소를 저장할 포인터
+    BinTreeNode *preturn = NULL;//찾는 노드의 주소를 반환할 포인터
+    int i = strlen(position) - 1;
 
     if (!(node = find_parent(tree, position)))
     {
@@ -69,7 +70,7 @@ BinTreeNode *getNode(BinTree *tree,char *position)
     }
     return (preturn);
 }
-
+//인자로 받은 트리의 특정 위치에 value값을 가지는 이진 노드를 하나 삽입하는 함수.
 int     putNode(BinTree *tree,char *position, char value)
 {
     BinTreeNode *node = NULL;
@@ -81,6 +82,7 @@ int     putNode(BinTree *tree,char *position, char value)
         return (FALSE);
     }
     str_len = strlen(position);
+    
     // position에 lLrR이 아닌 값이 들어왔을 경우의 예외처리를 이어서.
     if (!(position[str_len - 1] == 'l' || position[str_len - 1] == 'L'
         || position[str_len - 1] == 'r' || position[str_len - 1] == 'R'))
@@ -89,14 +91,11 @@ int     putNode(BinTree *tree,char *position, char value)
         return (FALSE);
     }
     // parent 노드의 leftchild에 노드를 넣는다.
-    BinTreeNode ele;
+    BinTreeNode ele;//value값을 넣어 만들 이진 노드
     memset(&ele, 0, sizeof(BinTreeNode));
     ele.data = value;
     if (position[str_len - 1] == 'l' || position[str_len - 1] == 'L')
     {
-
-        printf("node = %c\n", node->data);
-        //printf("node = %c\nnode->leftchild = %c\nele = %c\n", node->data, node->pLeftChild->data, ele.data);
         if (!insertLeftChildNodeBT(node, ele))
         {
             printf("[error] Failed to insert a node\n");
@@ -106,7 +105,6 @@ int     putNode(BinTree *tree,char *position, char value)
     // parent 노드의 rightchild에 노드를 넣는다.
     if (position[str_len - 1] == 'r' || position[str_len - 1] == 'R')
     {
-        //printf("node = %c\nnode->leftchild = %c\nele = %c", node->data, node->pLeftChild->data, ele.data);
         if (!insertRightChildNodeBT(node, ele))
         {
             printf("[error] Failed to insert a node\n");
@@ -116,74 +114,6 @@ int     putNode(BinTree *tree,char *position, char value)
     return (TRUE);
 }
 
-void    set_Nodes(BinTree *tree)
-{
-    char    *position;
-    char    value;
-    int     n = 0;
-    int     err_flag = 1;
-
-    if (!(position = (char *)malloc(sizeof(char) * 100)))
-    {
-        printf("[error] malloc failure : position.\n");
-        return ;
-    } 
-    while (1)
-    {
-        memset(position, 0, sizeof(char) * 100);
-        printf("=========================================================================\n");
-        printf("\n                        Node setting option\n\n");
-        printf("       1 : Insert Node       2 : Get Node         3 : Exit \n");
-        scanf(" %d", &n);
-        if (n == 3)
-        {
-            printf("Exit : Terminate the test program\n");
-            break;
-        }
-        switch (n)
-        {
-            case 1 :
-            {
-                printf("Enter the position and value.\n");
-                printf("Position : ");
-                scanf(" %s", position);
-                printf("value : ");
-                scanf(" %c", &value);
-                //printf("check position : %s\n", position);
-                err_flag = putNode(tree, position, value);
-                break;
-            }
-            case 2 :
-            {
-                BinTreeNode *node = NULL;
-                printf("Enter the position to find a node.\n");
-                printf("Position : ");
-                scanf(" %s", position);
-                if ((node = getNode(tree, position)) == NULL)
-                {
-                    err_flag = 0;
-                    break;
-                }
-                printf("Node : %c\n", node->data);
-                break;
-            }
-            default :
-            {
-                err_flag = 0;
-                printf("[error] Invalid Command\n");
-                break;
-            }
-        }
-        if (err_flag == 0)
-        {
-            printf("[error] Terminate the program.\n");
-            break;
-        }
-    }
-    free(position);
-}
-
-
 void    traversal(BinTree *tree)
 {
     int option = 0;
@@ -192,9 +122,8 @@ void    traversal(BinTree *tree)
     while (1)
     {
         printf("=========================================================================\n");
-        printf("\n                      Tree Traversal Option\n\n");
-        printf("   1 preorder    2 inorder    3 postorder   4 levelOrder    5 exit\n");
-        scanf(" %d", &option);
+        printf("Tree Traversal Option :\n   1 preorder    2 inorder    3 postorder   4 levelOrder    5 exit\n");
+        scanf("%d", &option);
         if (option == 5)
         {
             printf("Exit : Terminate the test program\n");
@@ -211,9 +140,99 @@ void    traversal(BinTree *tree)
             case 3 :
                 postorderTraversalBinTree(tree);
                 break;
-            // case 4 : 
-            //     levelOrderTraversalBinTree(tree);
-            //     break;
+            case 4 : 
+                levelOrderTraversalBinTree(tree);
+                break;
+            default :
+            {
+                err_flag = 0;
+                printf("[error] Invalid Command\n");
+                break;
+            }
+        }
+        if (err_flag == 0)
+            printf("[error] Terminate the program.\n");
+    }
+}
+
+void    set_Nodes(BinTree *tree)
+{
+    char    *position;
+    char    value;
+    int     n = 0;
+    int     err_flag = 1;
+
+    position = (char *)malloc(sizeof(char) * 100);
+    while (1)
+    {
+        memset(position, 0, sizeof(char) * 100);
+        printf("=========================================================================\n");
+        printf("\nNode setting option :\n");
+        printf("    1 : Insert Node    2 : Get Node    3 : Delete     4 : traversal     5 : Exit \n");
+        scanf(" %d", &n);
+        if (n == 5)
+        {
+            printf("Exit : Terminate the test program\n");
+            break;
+        }
+        switch (n)
+        {
+            case 1 :
+            {
+                if (deleteflag == 1)
+                {
+                    printf("already tree deleted\n");
+                    break;
+                }
+                printf("Enter the position and value.\n");
+                printf("Position : ");
+                scanf(" %s", position);
+                printf("value : ");
+                scanf(" %c", &value);
+                printf("check position : %s\n", position);
+                err_flag = putNode(tree, position, value);
+                break;
+            }
+            case 2 :
+            {
+                BinTreeNode *node = NULL;
+                if (deleteflag == 1)
+                {
+                    printf("already tree deleted\n");
+                    break;
+                }
+                printf("Enter the position to find a node.\n");
+                printf("Position : ");
+                scanf(" %s", position);
+                if ((node = getNode(tree, position)) == NULL)
+                {
+                    err_flag = 0;
+                    break;
+                }
+                printf("Node : %c\n", node->data);
+                break;
+            }
+            case 3 :
+            {
+                if (deleteflag == 1)
+                {
+                    printf("already tree deleted\n");
+                    break;
+                }
+                deleteflag = 1;
+                deleteBinTree(tree);
+                break;
+            }
+            case 4 :
+            {
+                if (deleteflag == 1)
+                {
+                    printf("already tree deleted\n");
+                    break;
+                }
+                traversal(tree);
+                break;
+            }
             default :
             {
                 err_flag = 0;
@@ -227,6 +246,8 @@ void    traversal(BinTree *tree)
             break;
         }
     }
+    free(position);
+    position = 0;
 }
 
 void    test()
@@ -240,9 +261,21 @@ void    test()
     tree = makeBinTree(root);
     printf(">> Set the Nodes\n\n");
     set_Nodes(tree);
-    printf(">> Traversal the Tree\n\n");
-    traversal(tree);
-    deleteBinTree(tree);
+    if (deleteflag == 0)
+    {
+        deleteBinTree(tree);
+    }
+}
+BinTreeNode *makeNewNodeBT(int value)
+{
+    BinTreeNode *pNode = (BinTreeNode *)malloc(sizeof(BinTreeNode));
+    if (pNode != NULL)
+    {
+        pNode->data = value;
+        pNode->pLeftChild = 0;
+        pNode->pRightChild = 0;
+    }
+    return (pNode);
 }
 
 int     main()
